refactor(orchestrator): hardening bundle + clippy 1.94 cleanup
Roadmap "Orchestrator Hardening" follow-ups from the Plan → Execute
review: dedup shared helpers, fix wrong project_root use, extend the
preinstall safety-net filter, regrain the build-lock key, and document
the UI-only contract on SuitePlanned / plan_event / DeviceId.

- merge_project_apps deduped; suite.rs calls orchestrator's copy
- parse_and_expand + run_flow_on_device use configured project_root
  instead of flow_dir fallback (so mixins resolve against the real root)
- slot-label renderers unified: shape_label + describe_slot pub in
  orchestrator; suite.rs no longer maintains a parallel copy
- build_install_matrix bounds-checks flow_idx (safe for external callers)
- device_matches_entry_constraints now filters on physical, name,
  playstore (was device_type only); accessibility_label left out
  since DeviceInfo has no counterpart
- project_lock keyed by (project_root, script_path) so monorepo apps
  sharing a root no longer serialise with each other; +1 test
- docs: SuitePlanned (UI-only strings contract), plan_event (take-once),
  compute_device_availability (eligibility, not parallel capacity),
  DeviceId (canonical "{platform}/{name}" scheme)
- device availability line simplified to "N device(s)"

Clippy 1.94 bumped strictness; rolling the mechanical cleanup into the
same commit so the tree stays -D warnings clean:

- .clone() dropped on Copy types (Bounds, Viewport, TreeStats)
- ReportAccumulator → #[derive(Default)]
- .last() → .rfind() on DoubleEndedIterators, .filter().next() → .find()
- redundant as u64 cast, redundant closure, collapsed if, trim before
  split_whitespace, unnecessary lifetime, doc list indentation
- mutex unwrap → expect with reason, single-arm match → if let
- type_complexity + too_many_arguments allows on installer internals

Verified: cargo test --workspace --lib = 1008 passed;
cargo clippy --workspace --lib -- -D warnings = clean.

+++ golem-devices/src/concurrency.rs
@@ -158,7 +158,7 @@ pub fn get_available_disk_mb() -> Result<u64> {
     }
 
     let stat = unsafe { stat.assume_init() };
-    let available_bytes = stat.f_bavail as u64 * stat.f_frsize as u64;
+    let available_bytes = stat.f_bavail as u64 * stat.f_frsize;
     Ok(available_bytes / (1024 * 1024))
 }
